refactor(category): order structure for extends
