+++ modules/@ergonode/categories/src/config/imports.js
@@ -17,10 +17,12 @@ export const Icons = {
     Templates: () => import('@Core/components/Icons/Menu/IconTemplates'),
 };
 
+// Extends
+
 export const Components = {
-    ProductFormCategory: () => import('@Categories/extends/components/Forms/ProductFormCategory'),
+    ProductFormCategory: () => import('@Categories/extends/product/components/Forms/ProductFormCategory'),
 };
 
 export const Store = {
-    Product: () => import('@Categories/extends/store/product').then(m => m.default || m),
+    Product: () => import('@Categories/extends/product/store').then(m => m.default || m),
 };
