+++ modules/@ergonode/categories/src/extends/categoryTree/components/SideBars/CategorySideBarElement.vue
