+++ modules/@ergonode/category-trees/src/components/Tabs/CategoryTreeDesignerTab.vue
@@ -73,7 +73,7 @@ export default {
             return [
                 {
                     title: 'Categories',
-                    component: () => import('@Categories/extends/components/VerticalTabs/CategoriesVerticalTab'),
+                    component: () => import('@Categories/extends/categoryTree/components/VerticalTabs/CategoriesVerticalTab'),
                     icon: () => import('@Trees/components/Icons/IconTree'),
                     props: {
                         isSelectLanguage: true,
